backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
opencv-python==4.9.0.80
numpy==1.26.4
torch>=2.2.0
ultralytics>=8.1.0
paddlepaddle==2.6.2
paddleocr==2.7.0.3
Pillow==10.2.0 